refactor: hover module for the language_server
